testing higher precision in wave_calib

+++ vcal/calib/calib_SPHERE.py
@@ -1520,7 +1520,7 @@ def calib(params_calib_name='VCAL_params_calib.json') -> None:
                     hdul = fits.open(inpath+wave_calib_list_ifs[ii])
                     wc_head = hdul[0].header
                     cube = hdul[0].data
-                    cube = np.array(cube, dtype=np.float32)
+                    cube = np.array(cube, dtype=np.float64)
                     ## MANUAL DARK SUBTRACTION
                     dit_wc = wc_head['HIERARCH ESO DET SEQ1 DIT']
                     for nn, fdit in enumerate(dit_ifs_flat_list):
